storage: reorganize TimeTrigger code
